requirements.txt
setuptools~=67.6.0
flask~=2.2.2
openai~=0.27.2
tiktoken~=0.3.2
bidict~=0.22.1
flask-socketio~=5.3.3
python-engineio~=4.4.0
python-socketio~=5.8.0
gevent==22.10.2
gevent-websocket==0.10.1
eventlet==0.33.3

websocket-client~=1.4.1